Refactor state interface construction
Move state metadata parsing and validation out of the converter, avoid regex parsing for JAX result leaf names, and isolate function state I/O planning from TranslationContext.
