+++ stablehlo_coreml/__init__.py
@@ -1,5 +1,6 @@
-from .converter import StateSpec, convert
+from .converter import convert
 from .passes.utils import DEFAULT_HLO_PIPELINE, register_optimizations
+from .state import StateSpec
 
 __version__ = "0.0.0"
 __all__ = ['DEFAULT_HLO_PIPELINE', 'StateSpec', 'convert', 'register_optimizations']
